Implements simple verifications on the controllers
Some of this verifications block certain endpoints that in a future
could be accessible by administrators. As that's not included on the
minimum requirements, there are blocked for now with a 403.

It's not very DRY, so in a future a lot of this validations could be
made on endpoint middlewares.